bugfix: 修复 filter_flows_can_create_task 创建资源实例属性不正确的问题
